Plugin wazo-snom with 10.1.39.11 firmware. Tested on D735 and D715. Change directory xivo-snom to wazo-snom
