+++ plugins/wazo-snom/10.1.20.0/entry.py
