+++ plugins/wazo-snom/10.1.20.0/pkgs/pkgs.db
