+++ plugins/wazo-snom/10.1.26.1/pkgs/pkgs.db
