+++ plugins/wazo-snom/10.1.26.1/plugin-info
