+++ plugins/wazo-snom/10.1.39.11/entry.py
@@ -8,8 +8,9 @@ execfile_('common.py', common_globals)
 
 MODELS = [
 	u'D375',
-	u'D715',
-	u'D725',
+	u'715',
+	u'D717',
+	u'725',
 	u'D735',
 	u'D745',
 	u'D765',
