+++ plugins/wazo-snom/8.7.5.35/pkgs/pkgs.db
