+++ plugins/wazo-snom/8.9.3.40/entry.py
