+++ plugins/wazo-snom/8.9.3.40/pkgs/pkgs.db
