+++ plugins/wazo-snom/8.9.3.60/pkgs/pkgs.db
