+++ plugins/wazo-snom/8.9.3.60/plugin-info
