+++ plugins/wazo-snom/8.9.3.80/pkgs/pkgs.db
