+++ plugins/wazo-snom/common/templates/D717.tpl
