--- plugins/xivo-snom/common/templates/370.tpl
@@ -1,1 +1,0 @@
-{% include 'base.tpl' -%}
