--- plugins/xivo-snom/common/templates/820.tpl
@@ -1,1 +1,0 @@
-{% include 'base.tpl' -%}
