--- plugins/xivo-snom/common/templates/821.tpl
@@ -1,1 +1,0 @@
-{% include 'base.tpl' -%}
