--- plugins/xivo-snom/common/templates/870.tpl
@@ -1,1 +1,0 @@
-{% include 'base.tpl' -%}
