--- plugins/xivo-snom/common/templates/D375.tpl
@@ -1,1 +1,0 @@
-{% include 'base.tpl' -%}
