--- plugins/xivo-snom/common/templates/D735.tpl
@@ -1,8 +1,0 @@
-{% extends 'base.tpl' -%}
-
-{% block gui_fkey1 %}{% endblock %}
-
-{% block settings_suffix %}
-<gui_fkey1 perm="R">F_ADR_BOOK</gui_fkey1>
-<gui_fkey4 perm="R">none</gui_fkey4>
-{% endblock %}
