--- plugins/xivo-snom/common/templates/MP.tpl
@@ -1,1 +1,0 @@
-{% include 'base.tpl' -%}
