--- plugins/xivo-snom/common/templates/PA1.tpl
@@ -1,1 +1,0 @@
-{% include 'base.tpl' -%}
