--- plugins/xivo-snom/common/templates/common/gui_lang.xml.tpl
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<gui-languages>
-  <language url="http://{{ ip }}:{{ http_port }}/i18n/gui_lang_DE.xml" name="Deutsch"/>
-  <language url="http://{{ ip }}:{{ http_port }}/i18n/gui_lang_EN.xml" name="English"/>
-  <language url="http://{{ ip }}:{{ http_port }}/i18n/gui_lang_ES.xml" name="Espanol"/>
-  <language url="http://{{ ip }}:{{ http_port }}/i18n/gui_lang_FR.xml" name="Francais"/>
-  <language url="http://{{ ip }}:{{ http_port }}/i18n/gui_lang_NL.xml" name="Dutch"/>
-</gui-languages>
