--- plugins/xivo-snom/common/templates/common/snom-model.xml.tpl.btpl
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<settings>
-  <phone-settings>
-    <firmware_status>http://{{ ip }}:{{ http_port }}/snom#MODEL#-firmware.xml</firmware_status>
-  </phone-settings>
-</settings>
