--- plugins/xivo-snom/common/templates/common/web_lang.xml.tpl
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<web-languages>
-  <language url="http://{{ ip }}:{{ http_port }}/i18n/web_lang_DE.xml" name="Deutsch"/>
-  <language url="http://{{ ip }}:{{ http_port }}/i18n/web_lang_EN.xml" name="English"/>
-  <language url="http://{{ ip }}:{{ http_port }}/i18n/web_lang_ES.xml" name="Espanol"/>
-  <language url="http://{{ ip }}:{{ http_port }}/i18n/web_lang_FR.xml" name="Francais"/>
-  <language url="http://{{ ip }}:{{ http_port }}/i18n/web_lang_NL.xml" name="Dutch"/>
-</web-languages>
