--- plugins/xivo-snom/common/templates/other/base.htm.tpl
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<setting-files>
-<!-- This file is necessary because of what I would call a priority bug in
-     the configuration files -->
-  <file url="http://{{ ip }}:{{ http_port }}/{{ XX_xml_filename }}"/>
-</setting-files>
